requirements.txt
google-auth-oauthlib>=1.0.0
google-auth-httplib2>=0.1.0
google-api-python-client>=2.100.0
numpy>=1.24.0
scikit-learn>=1.3.0
python-dateutil>=2.8.2
pytz>=2023.3
